Add files via upload
This GitHub repository contains the code and resources for My First HTML Project. It serves as an introduction to HTML, showcasing the fundamental concepts and structure of a basic web page.
Key Features:
HTML Markup: The repository includes the HTML files that form the core structure of the web page. It demonstrates the proper usage of HTML tags, such as headings, paragraphs, lists, links, and images.
Resources and Assets: The repository includes any additional resources, such as images, videos that are used in the project. This allows for a comprehensive overview of all the necessary files to recreate the web page.
Whether you're a beginner learning HTML or someone looking for a refresher, this repository provides a solid foundation to understand and practice the essentials of web development. Feel free to explore, learn, and use this repository as a starting point for your own HTML projects.